third_party/WebKit/LayoutTests/fast/dom/shadow/focus-navigation-with-multiple-shadow-roots-expected.txt

CONSOLE WARNING: line 87: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4668884095336448 for more details.
This tests that pressing Tab key should traverse into shadow DOM subtrees, and pressing Shift-Tab should reverse the order.

Should move from A to host in forward
PASS
Should move from host to host///younger-A in forward
PASS
Should move from host///younger-A to host//older-A in forward
PASS
Should move from host//older-A to host//older-B in forward
PASS
Should move from host//older-B to host///younger-B in forward
PASS
Should move from host///younger-B to light-child-B in forward
PASS
Should move from light-child-B to B in forward
PASS
Should move from B to light-child-B in backward
PASS
Should move from light-child-B to host///younger-B in backward
PASS
Should move from host///younger-B to host//older-B in backward
PASS
Should move from host//older-B to host//older-A in backward
PASS
Should move from host//older-A to host///younger-A in backward
PASS
Should move from host///younger-A to host in backward
PASS
Should move from host to A in backward
PASS
Test finished.
PASS successfullyParsed is true

TEST COMPLETE
